Added test for custom formats, fixed bugs
I expected this test to fail and require a certain amount of refactoring to
work correctly. I don't quite understand how Perl 6 decides which multi
candidates are available to the current block. I have been having a problem with
a custom extension being found in a thing I'm working on using this, so I
expected this test to fail in the same way. It didn't so I'm clearly not
understanding something about multi-dispatch that I need to learn.